move form logic up to parent
